requirements.txt
ollama
pyinstaller
sounddevice
soundfile
wxpython
numpy
openai-whisper
tts
